requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=2.1.0=pypi_0
accelerate=0.21.0=pypi_0
addict=2.4.0=pypi_0
aiohappyeyeballs=2.6.1=pypi_0
aiohttp=3.11.18=pypi_0
aiosignal=1.3.2=pypi_0
albucore=0.0.17=pypi_0
albumentations=1.4.16=pypi_0
annotated-types=0.7.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
asttokens=2.4.1=pypi_0
async-timeout=5.0.1=pypi_0
attrs=25.3.0=pypi_0
bayesian-optimization=2.0.3=pypi_0
beautifulsoup4=4.12.3=pypi_0
blobfile=3.0.0=pypi_0
braceexpand=0.1.7=pypi_0
ca-certificates=2025.2.25=h06a4308_0
cachetools=5.4.0=pypi_0
catalyst=22.4=pypi_0
certifi=2024.7.4=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
cloudpickle=3.0.0=pypi_0
cmake=3.30.2=pypi_0
colorama=0.4.6=pypi_0
concept-attention=0.1=dev_0
contourpy=1.3.0=pypi_0
convert=0.1.2=pypi_0
cubinlinker-cu11=0.3.0.post2=pypi_0
cuda-python=11.8.2=pypi_0
cudf-cu11=24.6.1=pypi_0
cuml-cu11=24.6.1=pypi_0
cupy-cuda11x=13.2.0=pypi_0
cycler=0.12.1=pypi_0
cython=3.0.10=pypi_0
dask=2024.5.1=pypi_0
dask-cuda=24.6.0=pypi_0
dask-cudf-cu11=24.6.1=pypi_0
dask-expr=1.1.1=pypi_0
datasets=3.5.0=pypi_0
diffusers=0.30.2=pypi_0
dill=0.3.8=pypi_0
distributed=2024.5.1=pypi_0
distributed-ucxx-cu11=0.38.0=pypi_0
dists-pytorch=0.1=pypi_0
einops=0.6.0=pypi_0
eval-type-backport=0.2.0=pypi_0
executing=2.1.0=pypi_0
facexlib=0.3.0=pypi_0
fastrlock=0.8.2=pypi_0
filelock=3.15.4=pypi_0
filterpy=1.4.5=pypi_0
fonttools=4.53.1=pypi_0
frozenlist=1.6.0=pypi_0
fsspec=2024.6.1=pypi_0
ftfy=6.2.3=pypi_0
future=1.0.0=pypi_0
fvcore=0.1.5.post20221221=pypi_0
gdown=4.7.1=pypi_0
gitdb=4.0.12=pypi_0
gitpython=3.1.45=pypi_0
grpcio=1.66.1=pypi_0
hf-xet=1.1.8=pypi_0
huggingface-hub=0.34.4=pypi_0
hydra-slayer=0.5.0=pypi_0
icecream=2.1.3=pypi_0
idna=3.7=pypi_0
imageio=2.35.1=pypi_0
imgaug=0.4.0=pypi_0
importlib-metadata=8.2.0=pypi_0
importlib-resources=6.4.4=pypi_0
iopath=0.1.10=pypi_0
jinja2=3.1.4=pypi_0
joblib=1.4.2=pypi_0
kiwisolver=1.4.5=pypi_0
kornia=0.7.3=pypi_0
kornia-rs=0.1.5=pypi_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libstdcxx-ng=9.1.0=hdf63c60_0
libucx-cu11=1.15.0.post1=pypi_0
lit=18.1.8=pypi_0
llvmlite=0.43.0=pypi_0
lmdb=1.5.1=pypi_0
locket=1.0.0=pypi_0
lpips=0.1.4=pypi_0
lxml=5.3.0=pypi_0
markdown=3.7=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.9.2=pypi_0
mdurl=0.1.2=pypi_0
mpmath=1.3.0=pypi_0
msgpack=1.0.8=pypi_0
multidict=6.4.3=pypi_0
multiprocess=0.70.16=pypi_0
muse=0.0.1=pypi_0
mypy-extensions=1.0.0=pypi_0
narwhals=2.7.0=pypi_0
ncurses=6.3=h7f8727e_2
networkx=3.2.1=pypi_0
numba=0.60.0=pypi_0
numpy=1.26.4=pypi_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cublas-cu12=12.8.4.1=pypi_0
nvidia-cuda-cupti-cu11=11.7.101=pypi_0
nvidia-cuda-cupti-cu12=12.8.90=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-nvrtc-cu12=12.8.93=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu12=12.8.90=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
nvidia-cudnn-cu12=9.10.2.21=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-cufft-cu12=11.3.3.83=pypi_0
nvidia-cufile-cu12=1.13.1.3=pypi_0
nvidia-curand-cu11=10.2.10.91=pypi_0
nvidia-curand-cu12=10.3.9.90=pypi_0
nvidia-cusolver-cu11=11.4.0.1=pypi_0
nvidia-cusolver-cu12=11.7.3.90=pypi_0
nvidia-cusparse-cu11=11.7.4.91=pypi_0
nvidia-cusparse-cu12=12.5.8.93=pypi_0
nvidia-cusparselt-cu12=0.7.1=pypi_0
nvidia-nccl-cu11=2.14.3=pypi_0
nvidia-nccl-cu12=2.27.3=pypi_0
nvidia-nvjitlink-cu12=12.8.93=pypi_0
nvidia-nvtx-cu11=11.7.91=pypi_0
nvidia-nvtx-cu12=12.8.90=pypi_0
nvtx=0.2.10=pypi_0
omegaconf=2.3.0=pypi_0
open-clip-torch=2.30.0=pypi_0
openai-clip=1.0.1=pypi_0
opencv-python=4.10.0.84=pypi_0
opencv-python-headless=4.10.0.84=pypi_0
openssl=1.1.1w=h7f8727e_0
packaging=24.1=pypi_0
pandas=2.2.2=pypi_0
partd=1.4.2=pypi_0
pillow=10.4.0=pypi_0
pip=24.0=py39h06a4308_0
platformdirs=4.3.6=pypi_0
plotly=6.3.1=pypi_0
portalocker=2.10.1=pypi_0
propcache=0.3.1=pypi_0
protobuf=5.28.2=pypi_0
psutil=6.0.0=pypi_0
ptxcompiler-cu11=0.8.1.post1=pypi_0
pyarrow=16.1.0=pypi_0
pycryptodomex=3.20.0=pypi_0
pydantic=2.9.2=pypi_0
pydantic-core=2.23.4=pypi_0
pygments=2.18.0=pypi_0
pylibraft-cu11=24.6.0=pypi_0
pynvml=11.4.1=pypi_0
pyparsing=3.1.4=pypi_0
pyre-extensions=0.0.23=pypi_0
pysocks=1.7.1=pypi_0
python=3.9.12=h12debd9_1
python-dateutil=2.9.0.post0=pypi_0
pytorch-warmup=0.1.1=pypi_0
pytorch-wavelets=1.3.0=pypi_0
pytz=2024.1=pypi_0
pywavelets=1.6.0=pypi_0
pyyaml=6.0.1=pypi_0
raft-dask-cu11=24.6.0=pypi_0
rapids-dask-dependency=24.6.0=pypi_0
rawpy=0.23.1=pypi_0
readline=8.1.2=h7f8727e_1
regex=2024.7.24=pypi_0
requests=2.32.3=pypi_0
rich=13.7.1=pypi_0
rmm-cu11=24.6.0=pypi_0
safetensors=0.4.4=pypi_0
scikit-image=0.24.0=pypi_0
scikit-learn=1.5.2=pypi_0
scipy=1.13.1=pypi_0
sentencepiece=0.2.0=pypi_0
sentry-sdk=2.40.0=pypi_0
setuptools=69.5.1=py39h06a4308_0
shapely=2.0.6=pypi_0
six=1.16.0=pypi_0
smmap=5.0.2=pypi_0
sortedcontainers=2.4.0=pypi_0
soupsieve=2.6=pypi_0
sqlite=3.38.5=hc218d9a_0
ssim-pil=1.0.14=pypi_0
submitit=1.5.1=pypi_0
sympy=1.14.0=pypi_0
tabulate=0.9.0=pypi_0
tblib=3.0.0=pypi_0
tensorboard=2.17.1=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorboardx=2.6.2.2=pypi_0
termcolor=2.4.0=pypi_0
threadpoolctl=3.5.0=pypi_0
tifffile=2024.8.30=pypi_0
timm=1.0.9=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.3=pypi_0
tomli=2.0.1=pypi_0
toolz=0.12.1=pypi_0
torch=2.8.0=pypi_0
torch-dct=0.1.6=pypi_0
torchdata=0.11.0=pypi_0
torchmetrics=0.10.3=pypi_0
torchtext=0.18.0=pypi_0
torchvision=0.23.0=pypi_0
tornado=6.4.1=pypi_0
tqdm=4.66.5=pypi_0
transformers=4.33.0=pypi_0
treelite=4.1.2=pypi_0
triton=3.4.0=pypi_0
typing-extensions=4.12.2=pypi_0
typing-inspect=0.9.0=pypi_0
tzdata=2024.1=pypi_0
ucx-py-cu11=0.38.0=pypi_0
ucxx-cu11=0.38.0=pypi_0
urllib3=2.2.2=pypi_0
wandb=0.22.2=pypi_0
wcwidth=0.2.13=pypi_0
webdataset=0.2.100=pypi_0
werkzeug=3.0.4=pypi_0
wget=3.2=pypi_0
wheel=0.43.0=py39h06a4308_0
xformers=0.0.32.post2=pypi_0
xxhash=3.5.0=pypi_0
xz=5.2.5=h7f8727e_1
yacs=0.1.8=pypi_0
yapf=0.40.2=pypi_0
yarl=1.20.0=pypi_0
zict=3.0.0=pypi_0
zipp=3.19.2=pypi_0
zlib=1.2.12=h7f8727e_2
